Actually working version
This version works correctly for the static images.
For this commit, created a function at the end, to loop through the images, and process them in a logical sequential steps. 
Also cleaned up the code, by removing obsoleted comments and most output.

Actual dev work included properly ordering the polygon creation, unwarping it, and adding it to the original image.