Got it working almost correctly across all frames in project video
Expanded the thresholding functions mostly
Split into :
Get Sobel for X
Get Sobel for Y
Get Magnitude of Sobel
Get Direction of Sobel

Scaling for sobel outputs was important and using 255 rather than 1 (as the image was consistently scaled to [0..255]
Take 32F as the input for Sobels rather than 64F or 8U

Made the windowing functions larger to pick up more points, and used less points in the determination of a window recenter. This was when performance was poor for the Sobels, so potentially change back to be stricter.

Added in zero array for the fits, in case there are no points to fit, to hopefully prevent errors